requirements_pylaia.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
docstring_parser==0.17.0
exceptiongroup==1.3.1
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
google==3.0.0
google-auth==2.43.0
google-genai==1.52.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
imageio==2.37.0
imagesize==1.4.1
Jinja2==3.1.3
jsonargparse==4.7.0
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.9.2
mdutils==1.6.0
mpmath==1.3.0
multidict==6.7.0
natsort==8.4.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==11.3.0
prettytable==3.11.0
propcache==0.4.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyDeprecate==0.3.1
pylaia==1.1.2
pyparsing==3.2.5
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textdistance==4.6.3
tifffile==2025.5.10
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.7.3
torchvision==0.14.1
tqdm==4.67.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
